chore(deletions) Move ScheduledDeletion models to deletions
Move the `ScheduledDeletion` model classes into `sentry.deletions` so
that all the deletion code is co-located.

Refs #77479
