+++ src/sentry/api/endpoints/api_application_details.py
@@ -10,8 +10,8 @@ from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import Endpoint, control_silo_endpoint
 from sentry.api.exceptions import ResourceDoesNotExist
 from sentry.api.serializers import serialize
+from sentry.deletions.models.scheduleddeletion import ScheduledDeletion
 from sentry.models.apiapplication import ApiApplication, ApiApplicationStatus
-from sentry.models.scheduledeletion import ScheduledDeletion
 
 
 class ApiApplicationSerializer(serializers.Serializer):
