+++ src/sentry/deletions/models/__init__.py
@@ -1,0 +1,11 @@
+from sentry.deletions.models.scheduleddeletion import (
+    RegionScheduledDeletion,
+    ScheduledDeletion,
+    get_regional_scheduled_deletion,
+)
+
+__all__ = (
+    "get_regional_scheduled_deletion",
+    "ScheduledDeletion",
+    "RegionScheduledDeletion",
+)
