+++ tests/sentry/api/endpoints/test_api_application_details.py
@@ -1,7 +1,7 @@
 from django.urls import reverse
 
+from sentry.deletions.models.scheduleddeletion import ScheduledDeletion
 from sentry.models.apiapplication import ApiApplication, ApiApplicationStatus
-from sentry.models.scheduledeletion import ScheduledDeletion
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.silo import control_silo_test
