+++ tests/sentry/integrations/api/endpoints/test_organization_integration_details.py
@@ -1,10 +1,10 @@
 from unittest.mock import patch
 
+from sentry.deletions.models.scheduleddeletion import ScheduledDeletion
 from sentry.integrations.base import IntegrationInstallation
 from sentry.integrations.models.integration import Integration
 from sentry.integrations.models.organization_integration import OrganizationIntegration
 from sentry.models.repository import Repository
-from sentry.models.scheduledeletion import ScheduledDeletion
 from sentry.shared_integrations.exceptions import ApiError, IntegrationError
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import APITestCase
