+++ tests/sentry/integrations/vercel/test_uninstall.py
@@ -2,10 +2,10 @@ import responses
 
 from fixtures.vercel import SECRET
 from sentry.constants import ObjectStatus
+from sentry.deletions.models.scheduleddeletion import ScheduledDeletion
 from sentry.integrations.models.integration import Integration
 from sentry.integrations.models.organization_integration import OrganizationIntegration
 from sentry.integrations.vercel import VercelClient
-from sentry.models.scheduledeletion import ScheduledDeletion
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.helpers import override_options
 from sentry.testutils.silo import control_silo_test
